requirements.txt
flask~=3.0.3
numpy~=1.26.4
Pillow~=10.3.0
torch~=2.2.2
unicorn
requests~=2.31.0
git+https://github.com/xinntao/Real-ESRGAN.git

